==> 1_preparacao.ipynb <==
# Preparação dos Dados

### Importação das Bibliotecas

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import warnings
warnings.filterwarnings('ignore')

import plotly.graph_objects as go
import plotly.express as px

# plot parameters
%matplotlib inline
plt.rcParams['font.size'] = 14
plt.rcParams['figure.figsize'] = (9, 5)
plt.rcParams['figure.facecolor'] = '#00000000'

### Leitura dos Dados

df = pd.read_csv("data/dados.csv")
df_base = df

def target(line):
    line = float(line)
    if line >= 0 :
        return 1    
    else :
        return 0

df['Target'] = df['Profit'].apply(lambda x: target(x))

df.head()

df.columns

dfx = df[['Order_Priority', 'Order_Quantity', 'Sales', 'Ship_Mode',
       'Profit', 'Customer_Name', 'Region', 'Customer_Segment',
       'Product_Category', 'Product_Sub-Category', 'Product_Container', 'Target']]

dfx.head()

dfx.to_csv('data/dados_21.csv', index=None)

from sklearn.model_selection import train_test_split
train, test = train_test_split(dfx, test_size=0.2)
train.to_csv('data/dados_2_train.csv', index=None)
test.to_csv('data/dados_2_test.csv', index=None)

df.head()

df.columns

df.dtypes

df.describe()

df.isnull().sum()

len(df.columns)

len(df)

### Tratamento das colunas tipo string para tipo numérico: 
1. As colunas tipo string no dataset selecionado apresenta uma natureza categórica, ou seja, não-ordinária.
2. Desta forma, para cada coluna e seus respectivos valores, são criadas colunas artificiais com o valor "1" ou "0" .

np.unique(df['Order_Priority'], return_counts = True)
#Categórica_Nominal And Categórica_Ordinal

### Tratamento para a coluna 'Order_Priority'
1. Criado para cada valor categórico, é criado uma funcao e em seguida aplicado no dataset
2. Este tratamento é similar para as demais colunas

def order_pri_A(line):
    if line == 'Critical':
        return 1
    else:
        return 0

def order_pri_B(line):
    if line == 'High':
        return 1
    else:
        return 0

def order_pri_C(line):
    if line == 'Low':
        return 1
    else:
        return 0

def order_pri_D(line):
    if line == 'Medium':
        return 1
    else:
        return 0

def order_pri_E(line):
    if line == 'Not Specified':
        return 1
    else:
        return 0

df['Critical_Ship_Mode'] = df['Order_Priority'].apply(lambda x: order_pri_A(x))
df['High_Ship_Mode'] = df['Order_Priority'].apply(lambda x: order_pri_B(x))
df['Low_Ship_Mode'] = df['Order_Priority'].apply(lambda x: order_pri_C(x))
df['Medium_Ship_Mode'] = df['Order_Priority'].apply(lambda x: order_pri_D(x))
df['NotSfd_Ship_Mode'] = df['Order_Priority'].apply(lambda x: order_pri_E(x))

### Tratamento para a coluna 'Ship_Mode'

np.unique(df['Ship_Mode'], return_counts = True)

def ship_m_R(line):
    if line == 'Regular Air':
        return 1
    else:
        return 0

def ship_m_E(line):
    if line == 'Express Air':
        return 1
    else: 
        return 0

df['Regular_Ship_Mode'] = df['Ship_Mode'].apply(lambda x: ship_m_R(x))

df['Express_Ship_Mode'] = df['Ship_Mode'].apply(lambda x: ship_m_E(x))

### Tratamento para as demais colunas

np.unique(df['Region'], return_counts = True)
#Categórica_Nominal

def reg_F(line):
    if line == 'Atlantic':
        return 1
    else: 
        return 0

def reg_G(line):
    if line == 'Northwest Territories':
        return 1
    else: 
        return 0

def reg_H(line):
    if line == 'Nunavut':
        return 1
    else: 
        return 0

def reg_I(line):
    if line == 'Ontario':
        return 1
    else: 
        return 0

def reg_J(line):
    if line == 'Prarie':
        return 1
    else: 
        return 0

def reg_K(line):
    if line == 'West':
        return 1
    else: 
        return 0

df['Atlantic_Region'] = df['Region'].apply(lambda x: reg_F(x))
df['NT_Region']       = df['Region'].apply(lambda x: reg_G(x))
df['Nunavut_Region']  = df['Region'].apply(lambda x: reg_H(x))
df['Ontario_Region']  = df['Region'].apply(lambda x: reg_I(x))
df['Prarie_Region']   = df['Region'].apply(lambda x: reg_J(x))
df['West_Region']    = df['Region'].apply(lambda  x: reg_K(x))

np.unique(df['Customer_Segment'], return_counts = True)

def cs_L(line):
    if line == 'Consumer':
        return 1
    else: 
        return 0

def cs_M(line):
    if line == 'Corporate':
        return 1
    else: 
        return 0

def cs_N(line):
    if line == 'Home Office':
        return 1
    else: 
        return 0

def cs_O(line):
    if line == 'Small Business':
        return 1
    else: 
        return 0

df['Consumer_Customer_Segment'] = df['Customer_Segment'].apply(lambda x: cs_L(x))
df['Corporate_Customer_Segment'] = df['Customer_Segment'].apply(lambda x: cs_M(x))
df['HO_Customer_Segment'] = df['Customer_Segment'].apply(lambda x: cs_N(x))
df['SB_Customer_Segment'] = df['Customer_Segment'].apply(lambda x: cs_O(x))

np.unique(df['Product_Category'], return_counts = True)
#Categórica_Nominal

def prod_R(line):
    if line == 'Furniture':
        return 1
    else: 
        return 0

def prod_S(line):
    if line == 'Office Supplies':
        return 1
    else: 
        return 0

def prod_T(line):
    if line == 'Technology':
        return 1
    else: 
        return 0

df['Furniture_Product_Category']  = df['Product_Category'].apply(lambda x: prod_R(x))
df['OF_Product_Category']         = df['Product_Category'].apply(lambda x: prod_S(x))
df['Technology_Product_Category'] = df['Product_Category'].apply(lambda x: prod_T(x))


np.unique(df['Product_Container'], return_counts = True)
#Categórica_Nominal

def pc_U(line):
    if line == 'Large Box':
        return 1
    else: 
        return 0

def pc_V(line):
    if line == 'Medium Box':
        return 1
    else: 
        return 0

def pc_X(line):
    if line == 'Small Box':
        return 1
    else: 
        return 0

df['LB_Product_Container'] = df['Product_Container'].apply(lambda x: pc_U(x))
df['MB_Product_Container'] = df['Product_Container'].apply(lambda x: pc_V(x))
df['SB_Product_Container'] = df['Product_Container'].apply(lambda x: pc_X(x))


df.columns

### Finalização da preparação do dataframe
1 - serão removidas as colunas nominais e a coluna 'Order_ID' pois a mesma representa o código da operação (não contribuindo para a fase de treinamento). O mesmo se aplica à coluna 'Customer_Name' e 'Product_Name' <br>
2 - colunas removidas: <br>
'Order_ID', 'Order_Priority', 'Ship_Mode', 'Customer_Name', 'Region', 'Customer_Segment',
       'Product_Category', 'Product_Sub-Category','Product_Name','Product_Container'


df_prepare = df[['Order_Quantity', 'Sales', 'Profit', 'Critical_Ship_Mode', 
                 'High_Ship_Mode', 'Low_Ship_Mode', 'Medium_Ship_Mode', 
                 'NotSfd_Ship_Mode', 'Regular_Ship_Mode', 'Express_Ship_Mode', 
                 'Atlantic_Region', 'NT_Region', 'Nunavut_Region', 'Ontario_Region', 
                 'Prarie_Region','West_Region', 'Consumer_Customer_Segment',
                 'Corporate_Customer_Segment', 'HO_Customer_Segment','SB_Customer_Segment', 
                 'Furniture_Product_Category','OF_Product_Category', 
                 'Technology_Product_Category','LB_Product_Container', 
                 'MB_Product_Container', 'SB_Product_Container']]

df_prepare.head()

### Reordenação da preparação do dataframe

df_prepare = df[['Order_Quantity', 'Critical_Ship_Mode', 
                 'High_Ship_Mode', 'Low_Ship_Mode', 'Medium_Ship_Mode', 
                 'NotSfd_Ship_Mode', 'Regular_Ship_Mode', 'Express_Ship_Mode', 
                 'Atlantic_Region', 'NT_Region', 'Nunavut_Region', 'Ontario_Region', 
                 'Prarie_Region','West_Region', 'Consumer_Customer_Segment',
                 'Corporate_Customer_Segment', 'HO_Customer_Segment','SB_Customer_Segment', 
                 'Furniture_Product_Category','OF_Product_Category', 
                 'Technology_Product_Category','LB_Product_Container', 
                 'MB_Product_Container', 'SB_Product_Container',
                 'Sales', 'Profit']]

df_prepare.head()

### Gravação da dataframe finalizado nesta etapa

df_prepare.to_csv('data/dados_prep.csv', index=None)

df_p2 = df[['Order_ID','Order_Priority','Order_Quantity', 'Critical_Ship_Mode', 
                 'High_Ship_Mode', 'Low_Ship_Mode', 'Medium_Ship_Mode', 
                 'NotSfd_Ship_Mode', 'Regular_Ship_Mode', 'Express_Ship_Mode', 
                 'Atlantic_Region', 'NT_Region', 'Nunavut_Region', 'Ontario_Region', 
                 'Prarie_Region','West_Region', 'Consumer_Customer_Segment',
                 'Corporate_Customer_Segment', 'HO_Customer_Segment','SB_Customer_Segment', 
                 'Furniture_Product_Category','OF_Product_Category', 
                 'Technology_Product_Category','LB_Product_Container', 
                 'MB_Product_Container', 'SB_Product_Container', 'Sales', 'Profit']]

df_p2.to_csv('data/dados_prep_index.csv', index=None)
