==> 4_modelo_aprendizagem_otimizacao.ipynb <==
# Preparação dos Dados

### Importação das Bibliotecas

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import warnings
warnings.filterwarnings('ignore')

import plotly.graph_objects as go
import plotly.express as px

# plot parameters
%matplotlib inline
plt.rcParams['font.size'] = 14
plt.rcParams['figure.figsize'] = (9, 5)
plt.rcParams['figure.facecolor'] = '#00000000'

from sklearn.model_selection import train_test_split
from sklearn.svm import SVR
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler

from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression

from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
import matplotlib.pyplot as plt
import plotly.express as px
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn import preprocessing
import pickle
from sklearn.model_selection import train_test_split
from sklearn.datasets import make_classification
from sklearn import svm
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, LinearSVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.neural_network import MLPClassifier

from sklearn.model_selection import cross_val_score

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import warnings
warnings.filterwarnings('ignore')

import plotly.graph_objects as go
import plotly.express as px

# plot parameters
%matplotlib inline
plt.rcParams['font.size'] = 14
plt.rcParams['figure.figsize'] = (9, 5)
plt.rcParams['figure.facecolor'] = '#00000000'

from sklearn.model_selection import train_test_split
from sklearn.svm import SVR
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler

from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression

from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
import matplotlib.pyplot as plt
import plotly.express as px
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn import preprocessing
import pickle
from sklearn.model_selection import train_test_split
from sklearn.datasets import make_classification
from sklearn import svm
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, LinearSVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.neural_network import MLPClassifier

from sklearn.model_selection import cross_val_score

from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC, LinearSVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.neural_network import MLPClassifier

from sklearn.model_selection import cross_val_score


### Leitura dos Dados

df_ini  = pd.read_csv("data/dados.csv")
df_prep = pd.read_csv("data/dados_prep.csv")
df_ind  = pd.read_csv("data/dados_prep_index.csv")

df_prep.columns

df_ini.columns

### Modelo de Predição para Order_Priority para o valor 'Not Specified''

df_esp     = df_ind.query(''' Order_Priority !='Not Specified' ''')

df_esp[['Order_Priority']].sample(5)

def order_pri(line):
    if line == 'Critical':
        return 1   
    elif line == 'High':
        return 2   
    elif line == 'Low':
        return 3   
    elif line == 'Medium':
        return 4   

df_esp['Type_Order_Priority'] = df_esp['Order_Priority'].apply(lambda x: order_pri(x))

df_esp.sample(5)

df_not_spec = df_ind.query(''' Order_Priority =='Not Specified' ''')

df_esp.columns

df = df_esp

X =  df.drop(['Order_ID','Profit', 'Order_Priority','Type_Order_Priority', ], axis = 1)
Y =  df['Type_Order_Priority']

seed = 1234
X_train, X_test, y_train, y_test = train_test_split(X,Y, test_size = 0.2, random_state = seed )

KNN_model = KNeighborsClassifier(n_neighbors=5)

KNN_model.fit(X_train, y_train)
KNN_prediction = KNN_model.predict(X_test)

print(accuracy_score(KNN_prediction, y_test))

print(classification_report(KNN_prediction, y_test))

### Execução com N-Modelos 

models = []

models.append(("LogisticRegression",LogisticRegression()))
models.append(("SVC",SVC()))
models.append(("LinearSVC",LinearSVC()))
models.append(("KNeighbors",KNeighborsClassifier()))
models.append(("DecisionTree",DecisionTreeClassifier()))
models.append(("RandomForest",RandomForestClassifier()))
rf2 = RandomForestClassifier(n_estimators=100, criterion='gini',
                                max_depth=10, random_state=0, max_features=None)
models.append(("RandomForest2",rf2))
models.append(("MLPClassifier",MLPClassifier(solver='lbfgs', random_state=0)))


results = []
names = []
for name,model in models:
    result = cross_val_score(model, X,Y,  cv=3)
    names.append(name)
    results.append(result)

for i in range(len(names)):
    print(names[i],results[i].mean())

### Execução com o Modelo que apresentou o melhor resultado

rf2 = RandomForestClassifier(n_estimators=100, criterion='gini',
                                max_depth=10, random_state=0, max_features=None)

rf2.fit(X_train, y_train)

rf2_prediction = rf2.fit(X_train, y_train).predict(X_test)

print(accuracy_score(rf2_prediction, y_test))

print(classification_report(rf2_prediction, y_test))

df_not_spec['Type_Order_Priority'] = df_not_spec['Order_Priority'].apply(lambda x: order_pri(x))

X =  df_not_spec.drop(['Order_ID','Profit', 'Order_Priority','Type_Order_Priority', ], axis = 1)

model = rf2.fit(X_train, y_train)

rf2_prediction = model.predict(X)

df_not_spec['Type_Order_Priority_Pred'] = rf2_prediction

df_not_spec.query(''' Type_Order_Priority_Pred == 4 ''')

df_not_spec.sample(10)

df = pd.DataFrame()
df     = pd.concat([df,df_not_spec]) 
df     = pd.concat([df,df_esp]) 

df.shape

df.columns

df.sample(5)

def ord_prix(v):
    if v == 1 : 
        return 1
    elif v == 2 :
        return 2
    elif v == 3 :
        return 3
    elif v == 4 :
        return 4
    else:
        return 4

df['Type_Order_Prior'] = df['Type_Order_Priority'].apply(lambda x: ord_prix(x))

df.sample(5)

df.columns

df = df[[ 'Order_Quantity', 'Critical_Ship_Mode',
       'High_Ship_Mode', 'Low_Ship_Mode', 'Medium_Ship_Mode',
       'NotSfd_Ship_Mode', 'Regular_Ship_Mode', 'Express_Ship_Mode',
       'Atlantic_Region', 'NT_Region', 'Nunavut_Region', 'Ontario_Region',
       'Prarie_Region', 'West_Region', 'Consumer_Customer_Segment',
       'Corporate_Customer_Segment', 'HO_Customer_Segment',
       'SB_Customer_Segment', 'Furniture_Product_Category',
       'OF_Product_Category', 'Technology_Product_Category',
       'LB_Product_Container', 'MB_Product_Container', 'SB_Product_Container',
       'Sales', 'Profit', 'Type_Order_Prior']]

df[['Profit']].sample(5)

def target(line):
    line = float(line)
    if line >= 0 :
        return 1    
    else :
        return 0

df['Target'] = df['Profit'].apply(lambda x: target(x))

X =  df.drop(['Profit', 'Target'], axis = 1)
Y =  df['Target']

models = []

models.append(("LogisticRegression",LogisticRegression()))
models.append(("SVC",SVC()))
models.append(("LinearSVC",LinearSVC()))
models.append(("KNeighbors",KNeighborsClassifier()))
models.append(("DecisionTree",DecisionTreeClassifier()))
models.append(("RandomForest",RandomForestClassifier()))
rf2 = RandomForestClassifier(n_estimators=100, criterion='gini',
                                max_depth=10, random_state=0, max_features=None)
models.append(("RandomForest2",rf2))
models.append(("MLPClassifier",MLPClassifier(solver='lbfgs', random_state=0)))


results = []
names = []
for name,model in models:
    result = cross_val_score(model, X,Y,  cv=3)
    names.append(name)
    results.append(result)

for i in range(len(names)):
    print(names[i],results[i].mean())

rf2 = RandomForestClassifier(n_estimators=100, criterion='gini',
                                max_depth=10, random_state=0, max_features=None)

rf2.fit(X_train, y_train)
rf2_prediction = rf2.fit(X_train, y_train).predict(X_test)
print(accuracy_score(rf2_prediction, y_test))
print(classification_report(rf2_prediction, y_test))
