==> 7_modelo_aprendizagem_h20.ipynb <==
import pandas as pd

import h2o

from h2o.automl import H2OAutoML

h2o.init()

df = h2o.import_file("data/dados_21.csv")

# Set response column as a factor
df['Target'] = df['Target'].asfactor()
response='Target'

# Split the dataset into train and test
train, test = df.split_frame(ratios = [.8], seed = 1234)

# Choose which columns to encode
encoded_columns = ["Order_Priority", "Ship_Mode", 
                   "Customer_Name", "Region", "Customer_Segment",
                   "Product_Category","Product_Sub-Category",
                   "Product_Container"]

# For k_fold strategy we need to provide fold column
fold_column = "kfold_column"
train[fold_column] = train.kfold_column(n_folds=5, seed=1234)

from h2o.estimators import H2OTargetEncoderEstimator
from h2o.estimators.gbm import H2OGradientBoostingEstimator

# Train a TE model
df_te = H2OTargetEncoderEstimator(fold_column=fold_column,
                                       data_leakage_handling="k_fold",
                                       blending=True,
                                       inflection_point=3,
                                       smoothing=10,
                                       noise=0.15,     # In general, the less data you have the more regularization you need
                                       seed=1234)

df_te.train(x=encoded_columns,
                 y=response,
                 training_frame=train)


train.head()

## New target encoded train and test sets
train_te = df_te.transform(frame=train, as_training=True)
test_te = df_te.transform(frame=test, noise=0)

train_te.head()

y

x_baseline = ['Order_Priority_te','Ship_Mode_te', 'Customer_Name_te',
              'Region_te', 'Customer_Segment_te', 'Product_Category_te',
               'Product_Sub-Category_te', 'Product_Container_te',
               'Order_Quantity','Sales','Target']


x = x_baseline
y = "Target"
x.remove(y)

# For binary classification, response should be a factor
train_te[y] = train_te[y].asfactor()
test_te[y] = test_te[y].asfactor()

# Run AutoML for 20 base models
aml = H2OAutoML(max_models=40, seed=1)
aml.train(x=x, y=y, training_frame=test_te)

# View the AutoML Leaderboard
lb = aml.leaderboard
lb.head(rows=lb.nrows)  # Print all rows instead of default (10 rows)

aml.leader

lb = h2o.automl.get_leaderboard(aml, extra_columns = "ALL")
lb

https://docs.h2o.ai/h2o/latest-stable/h2o-py/docs/intro.html

https://medium.com/h2o-ai-brasil/h2o-os-primeiros-passos-fae39077e028

https://docs.h2o.ai/h2o/latest-stable/h2o-docs/automl.html

https://docs.h2o.ai/h2o/latest-stable/h2o-docs/training-models.html
